Don't throw exception when "Add subscription" button is not present
This will happen when user with insufficient permissions access the
page. We rely on "Add subscription" button state as proxy for manifest
presence, but with recent changes this button may not be displayed at
all. In that case, exception would be thrown.

Also, wait a moment for button to appear. During testing it did happen
that page was "loaded", but button was not present yet.