Fix item selections on os table (#430)
The table has a filter mechanism that is causing problems with selecting an element in the row and causing failures to several test cases in test_operatingsystem.py. Need help in passing in the value for view.table.row() and self.parent.table.row() to get to the correct location.