Fix get_downloads_list and status locator.
here's why we used sleep here:
When we click on restart button then the process of generating and uploading a report gets started. There are two icons one for generating tab and other for uploading tab. They indicate users whether process of uploading or generating report is finished or in progress. So we were using those icon to wait till reports are generated and uploaded based on entity_item.status can result in uploading, generating and idle state. But the issue is that those icon don't change their state immediately after we start report generation process(clicking on restart button), which makes automation think that the process is finished(idle), which is not correct. That's why we have added sleep to wait for icons to change their state(generating or uploading). Also the process of generating/uploading reports takes some time either way, so I thought we can wait for few seconds.