Distinguish between top/bottom parameter new row
Add logic for distinguishing between parameters that add new rows to the
top of the parameters table vs the bottom of the parameters table. Add a
keyword argument to instantiation of the MatcherTable class in the
Ansible Variable view marking it as adding new parameter rows to the
bottom of the table.