Fix clean button locator in PF4 search. (#965)
As it was, we were getting an error when searching for the element,
getting False for is_displayed. Consequently, the clear button was
not clicked and a generic browser clear was called instead.
The text disappeared for a bit and then it got re-filled to the
search field.
It caused failures in tests such as
tests/foreman/ui/test_dashboard.py::test_positive_task_status
Changing type and locator fixes the issue.